Выполнены все правки Файл data разбит на составляющие по логике: добавлены файлы helpers.py , user_data_generator.py , user_api.py
Из conftest удалена фикстура user_data_delete_user в метод test_create_new_user добавлены строки кода

Дополнительно добавлен файл ingredient_data.py в котором содержатся хэши ингредиентов
