+++ src/data/ingredient_data.py
@@ -1,0 +1,7 @@
+class Ingredients:
+
+    one_correct_ingredients_data = {"ingredients": "61c0c5a71d1f82001bdaaa6e"}
+
+    two_correct_ingredients_data = {"ingredients": ["61c0c5a71d1f82001bdaaa6e", "61c0c5a71d1f82001bdaaa7a"]}
+
+    one_incorrect_ingredients_data = {"ingredients": "609646e4dc916e00276b28702"}
