+++ src/helpers.py
@@ -1,0 +1,7 @@
+import random
+import string
+
+def generate_random_string(length):
+    letters = string.ascii_lowercase
+    random_string = ''.join(random.choice(letters) for i in range(length))
+    return random_string
