+++ tests/tests_orders/test_order_create.py
@@ -3,20 +3,19 @@ import requests
 from conftest import user_data_with_delete_user
 from src.urls import Urls
 from src.messages import *
-from src.data import *
+from src.data.ingredient_data import Ingredients
+
 
 class TestOrderCreate:
-    @allure.title("Создание заказа с одни ингредиентом с неавторизованного пользователя")
+    @allure.title("Создание заказа с одним ингредиентом с неавторизованного пользователя")
     def test_create_order_without_login(self):
-        payload = {"ingredients": "61c0c5a71d1f82001bdaaa6e"}
-        responce = requests.post(Urls.url_order_create, json=payload)
+        responce = requests.post(Urls.url_order_create, json=Ingredients.one_correct_ingredients_data)
         assert responce.status_code == 200
         assert responce.json()["name"] == "Люминесцентный бургер"
 
     @allure.title("Создание заказа с несколькими ингредиентами с неавторизованного пользователя")
     def test_create_order_without_login_two_ingredients(self):
-        payload = {"ingredients": ["61c0c5a71d1f82001bdaaa6e", "61c0c5a71d1f82001bdaaa7a"]}
-        responce = requests.post(Urls.url_order_create, json=payload)
+        responce = requests.post(Urls.url_order_create, json=Ingredients.two_correct_ingredients_data)
         assert responce.status_code == 200
         assert responce.json()["name"] == "Люминесцентный астероидный бургер"
 
@@ -29,23 +28,20 @@ class TestOrderCreate:
 
     @allure.title("Создание заказа с некорректным хешем ингредиента")
     def test_create_order_with_incorrect_hash(self):
-        payload = {"ingredients": "609646e4dc916e00276b28702"}
-        responce = requests.post(Urls.url_order_create, json=payload)
+        responce = requests.post(Urls.url_order_create, json=Ingredients.one_incorrect_ingredients_data)
         assert responce.status_code == 500
 
-    @allure.title("Создание заказа с одни ингредиентом с авторизованного пользователя")
+    @allure.title("Создание заказа с одним ингредиентом с авторизованного пользователя")
     def test_create_order_with_login_user(self, user_data_with_delete_user):
-        payload = {"ingredients": "61c0c5a71d1f82001bdaaa6e"}
         headers = {"Authorization": user_data_with_delete_user["token"]}
-        response = requests.post(Urls.url_order_create,headers=headers, json=payload)
+        response = requests.post(Urls.url_order_create,headers=headers, json=Ingredients.one_correct_ingredients_data)
         assert response.status_code == 200
         assert response.json()["order"]["owner"]["email"] == user_data_with_delete_user["email"]
 
     @allure.title("Создание заказа с несколькими ингредиентами с авторизованного пользователя")
     def test_create_order_with_login_user_two_ingredients(self, user_data_with_delete_user):
-        payload = {"ingredients": ["61c0c5a71d1f82001bdaaa6e", "61c0c5a71d1f82001bdaaa7a"]}
         headers = {"Authorization": user_data_with_delete_user["token"]}
-        response = requests.post(Urls.url_order_create,headers=headers, json=payload)
+        response = requests.post(Urls.url_order_create,headers=headers, json=Ingredients.two_correct_ingredients_data)
         assert response.status_code == 200
         assert response.json()["order"]["price"] == 5130
